Add test for remove_aux_time()
move tests for cubes into their own file
