+++ tests/test_file_handling.py
@@ -67,32 +67,3 @@ def test_3d_spatial_weights(tmp_path):
     iris.save(domain, domain_file)
     expected_weights = np.array([24.0])
     assert helpers.nemo.spatial_weights(data, domain_file, "t") == expected_weights
-
-
-def test_load_input_cube():
-    src = "./tests/testdata/tos_nemo_all_mean_map.nc"
-    varname = "tos"
-    assert isinstance(helpers.cubes.load_input_cube(src, varname), Cube)
-
-
-def test_set_metadata():
-    cube = Cube([1])
-    cube.attributes = {
-        "description": None,
-        "interval_operation": None,
-        "interval_write": None,
-        "name": None,
-        "online_operation": None,
-    }
-    updated_cube = helpers.cubes.set_metadata(cube)
-    assert updated_cube.attributes == {"source": "EC-Earth 4", "Conventions": "CF-1.8"}
-    new_metadata = {
-        "title": "Title",
-        "comment": "Comment",
-        "diagnostic_type": "Type",
-        "source": "EC-Earth 4",
-        "Conventions": "CF-1.8",
-        "custom": "Custom",
-    }
-    updated_cube = helpers.cubes.set_metadata(updated_cube, **new_metadata)
-    assert updated_cube.attributes == new_metadata
